backend_py/requirements.txt
# backend_py/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfminer.six
pydantic
pymupdf
pillow
pytesseract
